Rename Detect class
