+++ src/LocaleServiceProvider.php
@@ -17,8 +17,8 @@ class LocaleServiceProvider extends ServiceProvider
             __DIR__ . '/config/locale.php' => config_path('thinktomorrow/locale.php'),
         ]);
 
-        $this->app->singleton(DetectLocaleAndScope::class, function ($app) {
-            return new DetectLocaleAndScope($app['request'], Config::from($this->getConfigValues()));
+        $this->app->singleton(Detect::class, function ($app) {
+            return new Detect($app['request'], Config::from($this->getConfigValues()));
         });
 
         $this->app->singleton(UrlParser::class, function ($app) {
@@ -36,7 +36,7 @@ class LocaleServiceProvider extends ServiceProvider
 
         $this->app->singleton(LocaleUrl::class, function ($app) {
             return new LocaleUrl(
-                $app['Thinktomorrow\Locale\DetectLocaleAndScope'],
+                $app['Thinktomorrow\Locale\Detect'],
                 $app['Thinktomorrow\Locale\Parsers\UrlParser'],
                 $app['Thinktomorrow\Locale\Parsers\RouteParser'],
                 Config::from($this->getConfigValues())
@@ -47,7 +47,7 @@ class LocaleServiceProvider extends ServiceProvider
          * Facade for getting current active scope
          */
         $this->app->singleton('tt-locale-scope', function ($app) {
-            return $app->make(DetectLocaleAndScope::class)->detectLocale()->getScope();
+            return $app->make(Detect::class)->detectLocale()->getScope();
         });
     }
