+++ tests/Unit/LocaleUrlTest.php
@@ -4,7 +4,7 @@ namespace Thinktomorrow\Locale\Tests\Unit;
 
 use Illuminate\Routing\Exceptions\UrlGenerationException;
 use Illuminate\Support\Facades\Route;
-use Thinktomorrow\Locale\DetectLocaleAndScope;
+use Thinktomorrow\Locale\Detect;
 use Thinktomorrow\Locale\Facades\LocaleUrlFacade;
 use Thinktomorrow\Locale\Tests\TestCase;
 
@@ -149,7 +149,7 @@ class LocaleUrlTest extends TestCase
         $this->get('http://foobar.com');
         $this->refreshBindings('Foobar');
 
-        Route::group(['prefix' => app(DetectLocaleAndScope::class)->detectLocale()->getScope()->activeSegment()], function () {
+        Route::group(['prefix' => app(Detect::class)->detectLocale()->getScope()->activeSegment()], function () {
             Route::get('/foo/bar/{color}', ['as' => 'foo.custom', 'uses' => function () {
             }]);
         });
